+++ tests/parser/test_parser.py
@@ -35,10 +35,10 @@ def test_parse_md_details(caplog):
     preprocessor = HTMLDocPreprocessor(docs_path, max_docs=max_docs)
 
     # Create an Parser and parse the md document
-    omni = Parser(
+    parser = Parser(
         structural=True, tabular=True, lingual=True, visual=True, pdf_path=pdf_path
     )
-    omni.apply(preprocessor, parallelism=PARALLEL)
+    parser.apply(preprocessor, parallelism=PARALLEL)
 
     # Grab the md document
     doc = session.query(Document).order_by(Document.name).all()[0]
@@ -117,10 +117,10 @@ def test_parse_md_paragraphs(caplog):
     preprocessor = HTMLDocPreprocessor(docs_path, max_docs=max_docs)
 
     # Create an Parser and parse the md document
-    omni = Parser(
+    parser = Parser(
         structural=True, tabular=True, lingual=True, visual=True, pdf_path=pdf_path
     )
-    omni.apply(preprocessor, parallelism=PARALLEL)
+    parser.apply(preprocessor, parallelism=PARALLEL)
 
     # Grab the document
     doc = session.query(Document).order_by(Document.name).all()[0]
@@ -211,8 +211,8 @@ def test_simple_tokenizer(caplog):
     # Preprocessor for the Docs
     preprocessor = HTMLDocPreprocessor(docs_path, max_docs=max_docs)
 
-    omni = Parser(structural=True, lingual=False, visual=True, pdf_path=pdf_path)
-    omni.apply(preprocessor, parallelism=PARALLEL)
+    parser = Parser(structural=True, lingual=False, visual=True, pdf_path=pdf_path)
+    parser.apply(preprocessor, parallelism=PARALLEL)
 
     doc = session.query(Document).order_by(Document.name).all()[1]
 
@@ -259,8 +259,8 @@ def test_parse_document_diseases(caplog):
     preprocessor = HTMLDocPreprocessor(docs_path, max_docs=max_docs)
 
     # Create an Parser and parse the diseases document
-    omni = Parser(structural=True, lingual=True, visual=True, pdf_path=pdf_path)
-    omni.apply(preprocessor, parallelism=PARALLEL)
+    parser = Parser(structural=True, lingual=True, visual=True, pdf_path=pdf_path)
+    parser.apply(preprocessor, parallelism=PARALLEL)
 
     # Grab the diseases document
     doc = session.query(Document).order_by(Document.name).all()[0]
@@ -374,8 +374,8 @@ def test_parse_style(caplog):
     preprocessor = HTMLDocPreprocessor(docs_path, max_docs=max_docs)
 
     # Create an Parser and parse the md document
-    omni = Parser(structural=True, lingual=True, visual=True, pdf_path=pdf_path)
-    omni.apply(preprocessor, parallelism=PARALLEL)
+    parser = Parser(structural=True, lingual=True, visual=True, pdf_path=pdf_path)
+    parser.apply(preprocessor, parallelism=PARALLEL)
 
     # Grab the document
     doc = session.query(Document).order_by(Document.name).all()[0]